requirements.txt
#Core requirements
pyusb